requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0
rank-bm25==0.2.2
nltk==3.9.1
numpy==1.26.4
torch==2.5.1
huggingface-hub==0.26.3
pydantic==2.10.3
python-multipart==0.0.20
httpx==0.28.1
